break up importer data component
