+++ src/components/importerData/ImportData.jsx
@@ -1,0 +1,86 @@
+import React, { Component } from 'react';
+import CSVReader from 'react-csv-reader';
+import CircularProgress from '@material-ui/core/CircularProgress';
+import { faFileCsv } from '@fortawesome/free-solid-svg-icons';
+import { ImportPreview, ImportMapper } from './';
+import { notify_error, notify_success, Icon } from '../../common.js';
+
+const sample_headers_from_csv = [
+  'First Name',
+  'Middle Initial',
+  'Last Name',
+  'Street #',
+  'Street Name',
+  'City Name',
+  'Postal Code',
+  'Position'
+];
+
+export default class ImporterData extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      loading: false,
+      data: null,
+      headers: []
+    };
+  }
+
+  preProcessError(e) {
+    notify_error(e, 'Failed to preprocess the import file.');
+  }
+
+  preProcess = async data => {
+    let headers = data.shift();
+    data.pop();
+
+    this.setState({ data, headers });
+  };
+
+  onHeadersSubmit = evt => {
+    evt.preventDefault();
+
+    this.setState({ loading: true });
+    // fake data loaded after 3 seconds
+    setTimeout(() => {
+      notify_success('Data has been imported.');
+      this.setState({ loading: false, headers: [] });
+    }, 3000);
+  };
+
+  render() {
+    if (this.state.loading) return <CircularProgress />;
+
+    if (!this.state.headers.length)
+      return (
+        <div>
+          <CSVReader
+            label="Data Importa"
+            onError={this.preProcessError}
+            onFileLoaded={this.preProcess}
+          />
+          <br />
+          <h3>Select a CSV file to get to the next menu!</h3>
+          (Also want the user to be able to drag&drop files.)
+        </div>
+      );
+
+    // convert headers to Select
+    let core_options = [];
+    sample_headers_from_csv.forEach(i =>
+      core_options.push({ value: i, label: i })
+    );
+
+    return (
+      <div>
+        <div style={{ display: 'flex' }}>
+          <h3>Import Data</h3> &nbsp;&nbsp;&nbsp;
+          <Icon icon={faFileCsv} size="3x" />
+        </div>
+        <ImportMapper core_options={core_options} />
+        <ImportPreview />
+      </div>
+    );
+  }
+}
