+++ src/components/importerData/ImportPreview.jsx
@@ -1,0 +1,119 @@
+import React from 'react';
+
+export const ImportPreview = () => (
+  <React.Fragment>
+    <h3>Sample Records based on selection</h3>
+    <div style={{ display: 'flex' }}>
+      <div>
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>Name:</div> <div>Joe Average</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>Street Address:</div>{' '}
+          <div>838 Wilshire Pl</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>City:</div> <div>Salt Lake City</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>State:</div> <div>NULL</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>Zip:</div> <div>84102</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>Country:</div> <div>NULL</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>Longitude:</div> <div>-111.8688189</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>Latitude</div> <div>40.7554569</div>
+        </div>
+      </div>
+
+      <div>
+        &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+      </div>
+
+      <div>
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>Joy B. Awesome</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>828 E Sego Ave</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>Salt Lake City</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>NULL</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>84102</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>NULL</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>-111.8677287</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>40.7550583</div>
+        </div>
+      </div>
+
+      <div>
+        &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+      </div>
+
+      <div>
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>Jake Abomination</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>857 Wilshire Pl</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>Salt Lake City</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>NULL</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>84102</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>NULL</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>-111.8688182</div>
+        </div>
+        <br />
+        <div style={{ display: 'flex' }}>
+          <div style={{ width: 150 }}>40.7554561</div>
+        </div>
+      </div>
+    </div>
+  </React.Fragment>
+);
