+++ src/components/importerData/index.js
@@ -1,0 +1,4 @@
+import { ImportMapper } from './ImportMapper';
+import { ImportPreview } from './ImportPreview';
+
+export { ImportMapper, ImportPreview };
